refactor : Players class from List to Set
