+++ src/main/java/model/user/Player.java
@@ -1,6 +1,5 @@
 package model.user;
 
-import java.util.Objects;
 import model.card.CardHand;
 import model.Result;
 
@@ -17,22 +16,5 @@ public class Player extends BlackJackPerson {
 
     public Result getResult() {
         return result;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Player player = (Player) o;
-        return result == player.result;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(result);
     }
 }
